* src/hsts.c: Fix -Wformat= warnings

+++ src/hsts.c
@@ -327,7 +327,8 @@ hsts_store_dump (hsts_store_t store, FILE *fp)
 
       if (fprintf (fp, "%s\t%d\t%d\t%lu\t%lu\n",
                    kh->host, kh->explicit_port, khi->include_subdomains,
-                   khi->created, khi->max_age) < 0)
+                   (unsigned long) khi->created,
+                   (unsigned long) khi->max_age) < 0)
         {
           logprintf (LOG_ALWAYS, "Could not write the HSTS database correctly.\n");
           break;
@@ -799,9 +800,9 @@ test_hsts_read_database (void)
       if (fp)
         {
           fputs ("# dummy comment\n", fp);
-          fprintf (fp, "foo.example.com\t0\t1\t%ld\t123\n",(long) created);
-          fprintf (fp, "bar.example.com\t0\t0\t%ld\t456\n", (long) created);
-          fprintf (fp, "test.example.com\t8080\t0\t%ld\t789\n", (long) created);
+          fprintf (fp, "foo.example.com\t0\t1\t%lu\t123\n",(unsigned long) created);
+          fprintf (fp, "bar.example.com\t0\t0\t%lu\t456\n", (unsigned long) created);
+          fprintf (fp, "test.example.com\t8080\t0\t%lu\t789\n", (unsigned long) created);
           fclose (fp);
 
           table = hsts_store_open (file);
